+++ src/DotNetHelper.Database/DataBaseAccessFactory.cs
@@ -570,6 +570,46 @@ namespace DotNetHelper.Database
             var parameters = ObjectToSql.BuildDbParameterList(instance, GetNewParameter, xmlSerializer, jsonSerializer, csvSerializer);
             return GetDataReader(sql, CommandType.Text, parameters);
         }
+
+
+
+        /// <summary>
+        /// Creates the specified SQL from the object then executes the sql and return the number of rows affected. 
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="instance">the object to create sql from </param>
+        /// <param name="actionType">type of sql to generate</param>
+        /// <param name="tableName">Table name to use when generating sql </param>
+        /// <param name="keyFields"></param>
+        /// <returns></returns>
+        public int Execute<T>(T instance, ActionType actionType, string tableName, params Expression<Func<T, object>>[] keyFields) where T : class
+        {
+            // var sql = (type.IsTypeAnonymousType() || type.IsTypeDynamic()) ? ObjectToSql.BuildQuery(actionType,instance, tableName ?? new SQLTable(DatabaseType, type).FullNameWithBrackets) : ObjectToSql.BuildQuery<T>( actionType, tableName ?? new SQLTable(DatabaseType, type).FullNameWithBrackets);
+            var sql = ObjectToSql.BuildQuery<T>(actionType, tableName ?? new SQLTable(DatabaseType, instance.GetType()).FullNameWithBrackets, keyFields);
+            var parameters = ObjectToSql.BuildDbParameterList(instance, GetNewParameter, null, null, null);
+            return ExecuteNonQuery(sql, CommandType.Text, parameters);
+        }
+
+
+        /// <summary>
+        /// Creates the specified SQL from the object then executes the sql and return the number of rows affected. 
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="instance">the object to create sql from </param>
+        /// <param name="actionType">type of sql to generate</param>
+        /// <param name="tableName">Table name to use when generating sql </param>
+        /// <param name="xmlSerializer">For when your storing values in the database as xml. This func will be invoke to serialize any property declarated with [SqlColumnAttribute(SerializableType = SerializableType.XML)]</param>
+        /// <param name="jsonSerializer">For when your storing values in the database as json. This func will be invoke to serialize any property declarated with [SqlColumnAttribute(SerializableType = SerializableType.JSON)]</param>
+        /// <param name="csvSerializer">For when your storing values in the database as csv. This func will be invoke to serialize any property declarated with [SqlColumnAttribute(SerializableType = SerializableType.CSV)]</param>
+        /// <param name="keyFields">Override attributes and specified which properties are keys from an expression</param>
+        /// <returns></returns>
+        public int Execute<T>(T instance, ActionType actionType, string tableName, Func<object, string> xmlSerializer, Func<object, string> jsonSerializer, Func<object, string> csvSerializer, params Expression<Func<T, object>>[] keyFields) where T : class
+        {
+            // var sql = (type.IsTypeAnonymousType() || type.IsTypeDynamic()) ? ObjectToSql.BuildQuery(actionType,instance, tableName ?? new SQLTable(DatabaseType, type).FullNameWithBrackets) : ObjectToSql.BuildQuery<T>( actionType, tableName ?? new SQLTable(DatabaseType, type).FullNameWithBrackets);
+            var sql = ObjectToSql.BuildQuery<T>(actionType, tableName ?? new SQLTable(DatabaseType, instance.GetType()).FullNameWithBrackets, keyFields);
+            var parameters = ObjectToSql.BuildDbParameterList(instance, GetNewParameter, xmlSerializer, jsonSerializer, csvSerializer);
+            return ExecuteNonQuery(sql, CommandType.Text, parameters);
+        }
     }
 }
 #endif