+++ tests/DotNetHelper.Database.Tests/BaseTest.cs
@@ -70,7 +70,6 @@ namespace DotNetHelper.Database.Tests
                 {
 
                 }
-
             });
         }